+++ src/main/types.ts
@@ -20,6 +20,7 @@ enum RecStatus {
   InvalidConfig,
   ReadyToRecord,
   FatalError,
+  Overruning,
 }
 
 /**